+++ internal/runtime/runtime.go
@@ -23,39 +23,45 @@ func code(isES6 bool) string {
 	// Note: The "__rest" function has a for-of loop which requires ES6, but
 	// transforming destructuring to ES5 isn't even supported so it's ok.
 	text := `
-		var __defineProperty = Object.defineProperty
-		var __hasOwnProperty = Object.prototype.hasOwnProperty
-		var __getOwnPropertySymbols = Object.getOwnPropertySymbols
-		var __getOwnPropertyDescriptor = Object.getOwnPropertyDescriptor
-		var __propertyIsEnumerable = Object.prototype.propertyIsEnumerable
+		var __create = Object.create
+		var __defProp = Object.defineProperty
+		var __getProtoOf = Object.getPrototypeOf
+		var __hasOwnProp = Object.prototype.hasOwnProperty
+		var __getOwnPropNames = Object.getOwnPropertyNames
+		var __getOwnPropDesc = Object.getOwnPropertyDescriptor
+		var __getOwnPropSymbols = Object.getOwnPropertySymbols
+		var __propIsEnum = Object.prototype.propertyIsEnumerable
 
 		export var __pow = Math.pow
 		export var __assign = Object.assign
+
+		// Tells importing modules that this can be considered an ES6 module
+		var __markAsModule = target => __defProp(target, '__esModule', { value: true })
 
 		// For object rest patterns
 		export var __restKey = key => typeof key === 'symbol' ? key : key + ''
 		export var __rest = (source, exclude) => {
 			var target = {}
 			for (var prop in source)
-				if (__hasOwnProperty.call(source, prop) && exclude.indexOf(prop) < 0)
+				if (__hasOwnProp.call(source, prop) && exclude.indexOf(prop) < 0)
 					target[prop] = source[prop]
-			if (source != null && __getOwnPropertySymbols)
+			if (source != null && __getOwnPropSymbols)
 	`
 
 	// Avoid "of" when not using ES6
 	if isES6 {
 		text += `
-				for (var prop of __getOwnPropertySymbols(source)) {
+				for (var prop of __getOwnPropSymbols(source)) {
 		`
 	} else {
 		text += `
-				for (var props = __getOwnPropertySymbols(source), i = 0, n = props.length, prop; i < n; i++) {
+				for (var props = __getOwnPropSymbols(source), i = 0, n = props.length, prop; i < n; i++) {
 					prop = props[i]
 		`
 	}
 
 	text += `
-					if (exclude.indexOf(prop) < 0 && __propertyIsEnumerable.call(source, prop))
+					if (exclude.indexOf(prop) < 0 && __propIsEnum.call(source, prop))
 						target[prop] = source[prop]
 				}
 			return target
@@ -71,13 +77,10 @@ func code(isES6 bool) string {
 		}
 
 		// Used to implement ES6 exports to CommonJS
-		var __markAsModule = target => {
-			return __defineProperty(target, '__esModule', { value: true })
-		}
 		export var __export = (target, all) => {
 			__markAsModule(target)
 			for (var name in all)
-				__defineProperty(target, name, { get: all[name], enumerable: true })
+				__defProp(target, name, { get: all[name], enumerable: true })
 		}
 		export var __exportStar = (target, module) => {
 			__markAsModule(target)
@@ -87,17 +90,17 @@ func code(isES6 bool) string {
 	// Avoid "let" when not using ES6
 	if isES6 {
 		text += `
-			for (let key in module)
-				if (!__hasOwnProperty.call(target, key) && key !== 'default')
-					__defineProperty(target, key, { get: () => module[key], enumerable: true })
+				for (let key of __getOwnPropNames(module))
+					if (!__hasOwnProp.call(target, key) && key !== 'default')
+						__defProp(target, key, { get: () => module[key], enumerable: __getOwnPropDesc(module, key).enumerable })
 		`
 	} else {
 		text += `
-			for (var key in module)
-				if (!__hasOwnProperty.call(target, key) && key !== 'default')
-					(k => {
-						__defineProperty(target, k, { get: () => module[k], enumerable: true })
-					})(key)
+				for (var keys = __getOwnPropNames(module), i = 0, n = keys.length, key; i < n; i++) {
+					key = keys[i]
+					if (!__hasOwnProp.call(target, key) && key !== 'default')
+						__defProp(target, key, { get: (k => module[k]).bind(null, key), enumerable: __getOwnPropDesc(module, key).enumerable })
+				}
 		`
 	}
 
@@ -110,7 +113,7 @@ func code(isES6 bool) string {
 			if (module && module.__esModule)
 				return module
 			return __exportStar(
-				__defineProperty({}, 'default', { value: module, enumerable: true }),
+				__defProp(__create(__getProtoOf(module)), 'default', { value: module, enumerable: true }),
 				module)
 		}
 
@@ -119,19 +122,19 @@ func code(isES6 bool) string {
 		// - kind === 1: method, parameter
 		// - kind === 2: field
 		export var __decorate = (decorators, target, key, kind) => {
-			var result = kind > 1 ? void 0 : kind ? __getOwnPropertyDescriptor(target, key) : target
+			var result = kind > 1 ? void 0 : kind ? __getOwnPropDesc(target, key) : target
 			for (var i = decorators.length - 1, decorator; i >= 0; i--)
 				if (decorator = decorators[i])
 					result = (kind ? decorator(target, key, result) : decorator(result)) || result
 			if (kind && result)
-				__defineProperty(target, key, result)
+				__defProp(target, key, result)
 			return result
 		}
 		export var __param = (index, decorator) => (target, key) => decorator(target, key, index)
 
 		// For class members
 		export var __publicField = (obj, key, value) => {
-			if (key in obj) return __defineProperty(obj, key, {enumerable: true, configurable: true, writable: true, value})
+			if (key in obj) return __defProp(obj, key, {enumerable: true, configurable: true, writable: true, value})
 			else return obj[key] = value
 		}
 		var __accessCheck = (obj, member, msg) => {